refactor: move log redaction to utils so telemetry stops depending on the engine
