+++ src/main/utils/log-redaction.js
